mike-paper-reviews-500/py_code/requirements.txt
python-docx>=0.8.11
lxml>=4.9.0
Pillow>=9.0.0
